# config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Documentação API PetMap",
    "description": "Esta documentação serve como guia para desenvolvedores entenderem como utilizar e integrar a API PetMap de forma eficiente, detalhando endpoints, parâmetros e métodos de autenticação.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira o token no formato: Bearer {token}"
    }
  },
  "paths": {
...
    "/api/commentary/create": {
      "post": {
        "tags": [
          "Commentary"
        ],
        "summary": "Registrar novo comentário em um pet",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "example": "any"
                },
                "petId": {
                  "example": "any"
                },
                "parentId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/commentary/list/{petId}": {
      "get": {
        "tags": [
          "Commentary"
        ],
        "summary": "Listar comentários de um pet específico",
        "description": "",
        "parameters": [
          {
            "name": "petId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/commentary/edit": {
      "put": {
        "tags": [
          "Commentary"
        ],
        "summary": "Editar texto de um comentário (somente o dono pode editar)",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "commentaryId": {
                  "example": "any"
                },
                "newText": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/commentary/remove": {
      "delete": {
        "tags": [
          "Commentary"
        ],
        "summary": "Remover comentário (apenas o dono do comentário pode remover)",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "commentaryId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/commentary/replies/{commentId}": {
      "get": {
        "tags": [
          "Commentary"
        ],
        "summary": "Listar todas as respostas recursivas de um comentário, incluindo o nome do usuário destinatário",
        "description": "",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/commentary/count-replies/{commentId}": {
      "get": {
        "tags": [
          "Commentary"
        ],
        "summary": "Contar o número total de respostas recursivas de um comentário",
        "description": "",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/commentary/{id}": {
      "get": {
        "tags": [
          "Commentary"
        ],
        "summary": "Buscar um comentário por ID (inclui respostas e pais)",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}